requirements/requirements-dev.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile --extra dev -o requirements/requirements-dev.txt pyproject.toml
asgiref==3.8.1
    # via
    #   django
    #   django-stubs
cfgv==3.4.0
    # via pre-commit
distlib==0.3.9
    # via virtualenv
django==5.1.3
    # via
    #   django-bank-app (pyproject.toml)
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework
django-stubs==5.1.1
    # via django-bank-app (pyproject.toml)
django-stubs-ext==5.1.1
    # via django-stubs
django-types==0.19.1
    # via django-bank-app (pyproject.toml)
djangorestframework==3.15.2
    # via django-bank-app (pyproject.toml)
factory-boy==3.3.1
    # via django-bank-app (pyproject.toml)
faker==33.0.0
    # via factory-boy
filelock==3.16.1
    # via virtualenv
identify==2.6.2
    # via pre-commit
mypy==1.13.0
    # via django-stubs
mypy-extensions==1.0.0
    # via mypy
nodeenv==1.9.1
    # via pre-commit
platformdirs==4.3.6
    # via virtualenv
pre-commit==4.0.1
    # via django-bank-app (pyproject.toml)
python-dateutil==2.9.0.post0
    # via faker
pyyaml==6.0.2
    # via pre-commit
ruff==0.7.3
    # via django-bank-app (pyproject.toml)
six==1.16.0
    # via python-dateutil
sqlparse==0.5.1
    # via django
types-psycopg2==2.9.21.20241019
    # via django-types
types-pyyaml==6.0.12.20240917
    # via django-stubs
typing-extensions==4.12.2
    # via
    #   django-stubs
    #   django-stubs-ext
    #   faker
    #   mypy
virtualenv==20.27.1
    # via pre-commit
